[aes] added CBF to encrypt / decrypt
