+++ crypto/aes/cfb.go
@@ -8,37 +8,12 @@ import (
 	"crypto/hmac"
 	"crypto/rand"
 	"crypto/sha256"
-	"encoding/json"
 	"errors"
 	"io"
 )
 
-type PlainText []byte
-
-type CipherText struct {
-	IV         []byte `json:"iv"`
-	CipherText []byte `json:"cipher_text"`
-	Mac        []byte `json:"mac"`
-	Version    uint8  `json:"v"`
-}
-
-// verify MAC of cipher text
-func (c CipherText) ValidMAC(s Secret) (bool, error) {
-	h := hmac.New(sha256.New, s[:])
-	if _, err := h.Write(c.CipherText); err != nil {
-		return false, err
-	}
-	return hmac.Equal(c.Mac, h.Sum(nil)), nil
-}
-
-func (c CipherText) Marshal() ([]byte, error) {
-	return json.Marshal(c)
-}
-
-type Secret [32]byte
-
 // encrypt a plain text with given secret
-func Encrypt(plainText PlainText, key Secret) (CipherText, error) {
+func CFBEncrypt(plainText PlainText, key Secret) (CipherText, error) {
 
 	block, err := aes.NewCipher(key[:])
 	if err != nil {
@@ -71,7 +46,7 @@ func Encrypt(plainText PlainText, key Secret) (CipherText, error) {
 }
 
 // decrypt a cipher text with given secret
-func Decrypt(cipherText CipherText, key Secret) (PlainText, error) {
+func CFBDecrypt(cipherText CipherText, key Secret) (PlainText, error) {
 
 	// create block
 	block, err := aes.NewCipher(key[:])
@@ -95,12 +70,4 @@ func Decrypt(cipherText CipherText, key Secret) (PlainText, error) {
 	stream.XORKeyStream(cc, cc)
 
 	return cc, nil
-}
-
-func Unmarshal(rawCipherText []byte) (CipherText, error) {
-	var ct CipherText
-	if err := json.Unmarshal(rawCipherText, &ct); err != nil {
-		return CipherText{}, err
-	}
-	return ct, nil
 }
