+++ crypto/aes/cfb_test.go
@@ -7,7 +7,7 @@ import (
 )
 
 //Test the encrypt and decrypt function in one batch
-func TestSuccessEncryptDecrypt(t *testing.T) {
+func CFBTestSuccessEncryptDecrypt(t *testing.T) {
 
 	secret := Secret{
 		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
@@ -18,11 +18,11 @@ func TestSuccessEncryptDecrypt(t *testing.T) {
 	value := []byte("I am the value")
 
 	//Encrypt
-	cipherText, e := Encrypt(value, secret)
+	cipherText, e := CFBEncrypt(value, secret)
 	require.Nil(t, e)
 
 	//Decrypt
-	res, err := Decrypt(cipherText, secret)
+	res, err := CFBDecrypt(cipherText, secret)
 	require.Nil(t, err)
 
 	//Decrypted value must match the given value
@@ -30,7 +30,7 @@ func TestSuccessEncryptDecrypt(t *testing.T) {
 
 }
 
-func TestFailedDecryption(t *testing.T) {
+func CFBTestFailedDecryption(t *testing.T) {
 
 	secret := Secret{
 		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
@@ -41,14 +41,14 @@ func TestFailedDecryption(t *testing.T) {
 	value := []byte("I am the plain text")
 
 	// encrypt
-	cipherText, e := Encrypt(value, secret)
+	cipherText, e := CFBEncrypt(value, secret)
 	require.Nil(t, e)
 
 	// change last byte to fail on decryption
 	secret[31] = 0x01
 
 	// decrypt
-	plainText, err := Decrypt(cipherText, secret)
+	plainText, err := CFBDecrypt(cipherText, secret)
 	require.Equal(t, PlainText{}, plainText)
 	require.EqualError(t, err, "invalid key - message authentication failed")
